run without print

+++ src/main/java/GApro/Fitness.java
@@ -12,10 +12,10 @@ package GApro;
 public class Fitness {
     private double fitnessScore;
     private String gtfStr;
-    private int generation;
+    private Long generation;
     
     
-    public Fitness(String gtfStr,int generation){
+    public Fitness(String gtfStr,Long generation){
         this.gtfStr=gtfStr;
         this.generation = generation;
         this.fitnessScore = (double)generation/10000;
@@ -29,7 +29,7 @@ public class Fitness {
         return gtfStr;
     }
 
-    public int getGeneration() {
+    public Long getGeneration() {
         return generation;
     }
